third_party/WebKit/LayoutTests/media/video-no-autoplay.html

<!DOCTYPE html>
<title>Test that play event does not fire when "src" set with no autoplay attribute.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<video></video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");
    assert_true(video.paused);

    video.onplay = t.unreached_func();

    video.oncanplaythrough = t.step_func(function() {
        setTimeout(t.step_func_done(function() {
            assert_true(video.paused);
        }), 500);
    });

    video.src = "content/test.ogv";
});
</script>
